Resume_JD_Agent_Backend/requirement.txt
flask
flask-cors
dotenv
langchain
langchain_core
langgraph
langchain_mistralai
langchain_ollama  
psycopg2-binary
langgraph
